requirements.txt
face-recognition==1.3.0
opencv-python==4.7.0.72
streamlit==1.22.0
mysql-connector-python==8.0.33
matplotlib==3.7.1
pandas==2.0.2
numpy==1.24.3
python-dotenv==1.0.0
pickle5==0.0.12